VBU-21 Check if index exists before deleting

+++ src/Services/ElasticaService.php
@@ -53,7 +53,10 @@ class ElasticaService
 
     public function reindex()
     {
-        $this->index->delete();
+        if ($this->index->exists()) {
+            $this->index->delete();
+        }
+
         $this->index->create();
 
         $documents = [];
